Fix stuttering by writing to back buffer in halves, increasing baud
The Arduino serial buffer is only 64 bytes long on a 32u4. Thus, when
receiving a 122-byte-long message, if there's any delay in reading out
into our own buffer, the Arduino buffer can fill, messing things up
significantly. The ability to write just half of the display (a
62-byte-long message) allows us to fit within the 64 bytes and not
have to win a race. We only add a command to do it into the back
buffer so that the intermediate state isn't seen.